requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6
aiosignal==1.3.1; python_full_version >= '3.7.0'
async-timeout==4.0.3; python_full_version >= '3.7.0'
attrs==23.1.0; python_full_version >= '3.7.0'
beautifulsoup4==4.12.2
cachetools==5.3.2
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
frozenlist==1.4.0; python_version >= '3.8'
idna==3.4; python_version >= '3.5'
lxml==4.9.3
multidict==6.0.4; python_full_version >= '3.7.0'
python-dotenv==1.0.0
soupsieve==2.5; python_version >= '3.8'
yarl==1.9.2; python_full_version >= '3.7.0'
